requirements.txt
flask==3.1.0
msal==1.31.1
python-dotenv==1.0.1
dnspython==2.7.0
requests==2.32.3
